This program sends messages using UDP broadcast.
When a message is sent to the broadcast address, all peers set to the broadcast address receive the message simultaneously.

Therefore, when a message is sent to the broadcast address, all peers on the network receive the message simultaneously.

In this way, messages can be delivered to all peers on the network. However, since it is delivered to all networks connected to the L2 segment, it puts a heavy load on the switch, ultimately leading to reduced efficiency.

Therefore, it is better to use broadcast only when necessary.